# docs/api/dashboard.openapi.yaml
openapi: "3.1.0"
info:
  title: msg2agent Dashboard API
  version: v1
  description: |
    REST API exposed by the msg2agent dashboard server.
    All `/api/dashboard/*` endpoints require a valid JWT obtained via the
    PKCE OAuth2 flow (see Auth section). Endpoints `/health`, `/version`, and
    `/api/email/resend` are public (no auth required).
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MeResponse:
      type: object
      required: [id, name, email, email_verified, plan, billing_status, created_at, quota]
      properties:
        id:
          type: string
          example: "ten_01j9abc123"
        name:
          type: string
          example: "Alice Smith"
        email:
          type: string
          format: email
          example: "alice@example.com"
        email_verified:
          type: boolean
          example: true
        plan:
          type: string
          enum: [free, starter, team]
          example: "starter"
        billing_status:
          type: string
          example: "active"
        current_period_end:
          type: string
          format: date-time
          nullable: true
          example: "2026-06-01T00:00:00Z"
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        quota:
          $ref: "#/components/schemas/QuotaConfig"
        did:
          type: string
          nullable: true
          description: Decentralized identifier (DID) derived from the tenant's DID seed.
          example: "did:key:z6MkhaXgBZDvotDkL5257faiztiGiC2QtKLGpbnnEGta2doK"
        signing_public_key:
          type: string
          nullable: true
          description: Base64-encoded Ed25519 signing public key.
          example: "MCowBQYDK2VwAyEA..."
        encryption_public_key:
          type: string
          nullable: true
          description: Base64-encoded X25519 encryption public key.
          example: "MCowBQYDK2VuAyEA..."
...
security:
  - bearerAuth: []
paths:
...
  /api/dashboard/me:
    get:
      summary: Current tenant identity
      description: |
        Returns full account information for the authenticated tenant,
        including plan, billing status, quota limits, and optional DID/key
        data when a DID seed is configured.
      operationId: getMe
      tags: [account]
      responses:
        "200":
          description: Tenant identity and quota.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/dashboard/profile:
    get:
      summary: Get profile
      description: Returns a subset of the tenant's profile (without plan/quota details).
      operationId: getProfile
      tags: [account]
      responses:
        "200":
          description: Profile data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update profile
      description: Updates the tenant's display name. Name must be 1–128 characters.
      operationId: patchProfile
      tags: [account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 128
                  example: "Alice Smith"
      responses:
        "200":
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "400":
          description: Validation error (empty name, name too long, invalid JSON).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/dashboard/keys:
    get:
      summary: List API keys
      description: |
        Returns a paginated list of all API keys for the tenant.
        Supports `?limit=` (default 100, max 500) and `?offset=` query parameters.
      operationId: listApiKeys
      tags: [api-keys]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
          example: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          example: 0
      responses:
        "200":
          description: Paginated list of API keys.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageApiKey"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create API key
      description: |
        Creates a new API key for the tenant. The full key value is returned
        only in this response and cannot be retrieved again. Label defaults to
        `"API Key"` when omitted and must be at most 64 characters.
        Rate-limited per tenant.
      operationId: createApiKey
      tags: [api-keys]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  maxLength: 64
                  example: "CI pipeline"
      responses:
        "201":
          description: API key created. Store the `key` value — it will not be shown again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateKeyResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying.
...
  /api/dashboard/usage:
    get:
      summary: Usage aggregates
      description: |
        Returns paginated usage rows (event counts per period) for the tenant.
        Filter by `?period=YYYY-MM` to get a specific month. Supports
        `?limit=` and `?offset=` pagination parameters.
      operationId: getUsage
      tags: [usage]
      parameters:
        - name: period
          in: query
          description: "Month filter in YYYY-MM format."
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
          example: "2026-05"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
          example: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          example: 0
      responses:
        "200":
          description: Paginated usage rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageUsageRow"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/dashboard/usage.csv:
    get:
      summary: Export usage as CSV
      description: |
        Downloads usage aggregates as a CSV file (`period,event,count`).
        Supports `?period=YYYY-MM` filter.
      operationId: getUsageCsv
      tags: [usage]
      parameters:
        - name: period
          in: query
          description: "Month filter in YYYY-MM format."
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
          example: "2026-05"
      responses:
        "200":
          description: CSV file attachment.
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="usage.csv"'
          content:
            text/csv:
              schema:
                type: string
              example: |
                period,event,count
                2026-05,messages,234
                2026-05,tool_calls,87
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/dashboard/usage/by-tool:
    get:
      summary: Usage breakdown by tool
      description: |
        Returns a paginated breakdown of tool call counts for the tenant,
        optionally filtered by `?period=YYYY-MM`. Returns 501 if the admin
        store is not configured.
      operationId: getUsageByTool
      tags: [usage]
      parameters:
        - name: period
          in: query
          description: "Month filter in YYYY-MM format."
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
          example: "2026-05"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated tool usage rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageToolUsageRow"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Admin store not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/dashboard/audit/verify:
    get:
      summary: Verify audit chain integrity
      description: |
        Walks the cryptographic chain of audit events for the authenticated
        tenant and returns verification results. Returns 501 if the admin
        store is not configured.
      operationId: getAuditVerify
      tags: [audit]
      responses:
        "200":
          description: Audit chain verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditVerifyResponse"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Admin store not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/dashboard/audit/events:
    get:
      summary: List audit events
      description: |
        Returns paginated audit events for the authenticated tenant.
        Supports `?period=YYYY-MM` to filter by month and `?limit=` (max 1000).
        Returns 501 if the admin store is not configured.
      operationId: listAuditEvents
      tags: [audit]
      parameters:
        - name: period
          in: query
          description: "Month filter in YYYY-MM format."
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
          example: "2026-05"
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
            maximum: 1000
          example: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated audit event list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAuditEventItem"
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Admin store not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...